osint_bot/requirements.txt
aiogram>=3.3.0
telethon>=1.33.1
aiohttp>=3.9.1
beautifulsoup4>=4.12.2
python-dotenv>=1.0.0
